moved highjack_transition_rando back
